Refactored Jarvis classes and gui. Changed tts driver to espeak for linux
