+++ GUI/.env
@@ -1,6 +1,15 @@
-USER = "Dhruv"
-BOT = "JARVIS"
+USER=Dhruv
+BOT=JARVIS
 EMAIL = ""
 PASSWORD = ""
-NEWS_API_KEY = ""
-OPENWEATHER_APP_ID = ""
+
+IP_ADDR_API_URL=https://api.ipify.org
+NEWS_FETCH_API_URL=https://newsapi.org/v2/top-headlines
+NEWS_FETCH_API_KEY=abcdedfghijklmnopqrstuvwxyz
+WEATHER_FORECAST_API_URL=https://api.openweathermap.org/data/2.5/weather
+WEATHER_FORECAST_API_KEY=abcdedfghijklmnopqrstuvwxyz 
+
+SMTP_URL=smtp.gmail.com
+SMTP_PORT=587
+
+
