--- GUI/conv.py
@@ -1,5 +1,0 @@
-random_text = [
-    "Cool, I'm on it sir.",
-    "Okay sir, I'm working on it.",
-    "Just a second sir.",
-]
