+++ requirements.txt
@@ -1,14 +1,60 @@
+aiohappyeyeballs==2.3.5
+aiohttp==3.10.3
+aiosignal==1.3.1
+asyncio==3.4.3
+attrs==24.2.0
+beautifulsoup4==4.12.3
+blinker==1.8.2
+certifi==2024.7.4
+cffi==1.17.0
+charset-normalizer==3.3.2
+cinemagoer==2023.5.1
+click==8.1.7
+disutils==1.4.32.post2
+docutils==0.21.2
+Flask==3.0.3
+frozenlist==1.4.1
+greenlet==3.0.3
+idna==3.7
 IMDbPY==2022.7.9
-IMDbPY==2022.7.9
+itsdangerous==2.2.0
+jaraco.context==5.3.0
+Jinja2==3.1.4
 keyboard==0.13.5
-Kivy==2.2.1
+Kivy==2.3.0
+Kivy-examples==2.3.0
+Kivy-Garden==0.1.5
+lxml==5.3.0
+MarkupSafe==2.1.5
+more-itertools==10.4.0
+MouseInfo==0.1.3
+multidict==6.0.5
 numpy==1.26.3
+pillow==10.4.0
+PyAudio==0.2.14
 PyAutoGUI==0.9.54
+pycparser==2.22
+PyGetWindow==0.0.9
+Pygments==2.18.0
+PyMsgBox==1.0.9
+pyperclip==1.9.0
+PyRect==0.2.0
+PyScreeze==0.1.30
 python-decouple==3.8
+python3-xlib==0.15
 pyttsx3==2.90
+pytweening==1.2.0
 pywhatkit==5.4
-Requests==2.31.0
+requests==2.31.0
+setuptools==72.1.0
 sounddevice==0.4.6
+soupsieve==2.5
 SpeechRecognition==3.10.0
+SQLAlchemy==2.0.32
+typing_extensions==4.12.2
+urllib3==2.2.2
+Werkzeug==3.0.3
 wikipedia==1.4.0
 wolframalpha==5.0.0
+xmltodict==0.13.0
+yarl==1.9.4
